Seperate functions from Pages/Edit into several files
