+++ src/Pages/Edit/Functions/onBackspacePressed.js
@@ -1,0 +1,3 @@
+const onBackspacePressed = (event, story, setStory) => {}
+
+export default onBackspacePressed
